fix(v1): datas.tsを作成し、データをこのファイルで一元管理するように変更
